requirements.txt
# Core Bot Framework
python-telegram-bot>=20.0
python-dotenv>=1.0.0

# LLM Integration
httpx>=0.24.0  # HTTP client with retry support

# Utilities
pyyaml>=6.0

